#51 move main components to the root dir
