+++ src/app/js/options/options.jsx
@@ -2,8 +2,8 @@ import "./options.scss";
 
 import React, { Component } from "react";
 
-import { CategorySelector } from "../category-selector/category-selector";
-import { Header } from "../header/header";
+import { CategorySelector } from "./components/category-selector/category-selector";
+import { Header } from "./components/header/header";
 
 export class Options extends Component {
     render() {
