+++ src/app/js/popup/index.jsx
@@ -1,7 +1,7 @@
 import "./index.scss";
 import "app/js/common/utils/axios.config";
 
-import { Popup } from "./components/popup/popup";
+import { Popup } from "./popup";
 import React from "react";
 import ReactDOM from "react-dom";
