+++ src/app/js/popup/popup.scss
